+++ benches/hnefatafl_rs.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use criterion::{Criterion, criterion_group, criterion_main};
+use criterion::{criterion_group, criterion_main, Criterion};
 
 use hnefatafl_copenhagen::hnefatafl_rs;