+++ src/bin/hnefatafl-client.rs
@@ -15,37 +15,37 @@ use std::{
 use std::path::PathBuf;
 
 use chrono::{Local, Utc};
-use clap::{Parser, command};
+use clap::{command, Parser};
 use env_logger::Builder;
 use futures::executor;
 use hnefatafl_copenhagen::{
-    HOME, VERSION_ID,
     color::Color,
     draw::Draw,
     game::Game,
-    glicko::{CONFIDENCE_INTERVAL_95, Rating},
+    glicko::{Rating, CONFIDENCE_INTERVAL_95},
     handle_error,
-    play::{BOARD_LETTERS, Vertex},
+    play::{Vertex, BOARD_LETTERS},
     rating::Rated,
     role::Role,
     server_game::{ServerGameLight, ServerGamesLight},
     space::Space,
     status::Status,
     time::{Time, TimeSettings},
+    HOME, VERSION_ID,
 };
 use iced::{
-    Element, Event, Subscription,
     alignment::{Horizontal, Vertical},
     event,
     font::Font,
     futures::{SinkExt, Stream},
     stream,
     widget::{
-        Column, Container, Row, Scrollable, button, checkbox, column, container, radio, row,
-        scrollable, text, text_input, tooltip,
+        button, checkbox, column, container, radio, row, scrollable, text, text_input, tooltip,
+        Column, Container, Row, Scrollable,
     },
+    Element, Event, Subscription,
 };
-use log::{LevelFilter, debug, error, info, trace};
+use log::{debug, error, info, trace, LevelFilter};
 
 const PORT: &str = ":49152";
 const PADDING: u16 = 10;
@@ -1318,10 +1318,8 @@ impl Client {
                     Status::WhiteWins => user_area = user_area.push(text("Defender Wins!")),
                 }
 
-                let spectator = column![
-                    text(format!("👥 ({})", self.spectators.len()))
-                        .shaping(text::Shaping::Advanced),
-                ];
+                let spectator = column![text(format!("👥 ({})", self.spectators.len()))
+                    .shaping(text::Shaping::Advanced),];
                 user_area = user_area.push(tooltip(
                     spectator,
                     container(spectators)
@@ -1361,9 +1359,9 @@ impl Client {
 
                 let leave = button("Leave").on_press(Message::Leave);
 
-                let mut time = row![
-                    checkbox("timed ", self.timed.clone().into()).on_toggle(Message::TimeCheckbox)
-                ];
+                let mut time =
+                    row![checkbox("timed ", self.timed.clone().into())
+                        .on_toggle(Message::TimeCheckbox)];
 
                 if self.timed.0.is_some() {
                     time = time.push(text("minutes"));
@@ -1593,14 +1591,12 @@ fn pass_messages() -> impl Stream<Item = Message> {
 
         let (tx, rx) = mpsc::channel();
         let _ = sender.send(Message::TcpConnected(tx)).await;
-        thread::spawn(move || {
-            loop {
-                let message = handle_error(rx.recv());
-                let message_trim = message.trim();
-                debug!("<- {message_trim}");
+        thread::spawn(move || loop {
+            let message = handle_error(rx.recv());
+            let message_trim = message.trim();
+            debug!("<- {message_trim}");
 
-                handle_error(tcp_stream.write_all(message.as_bytes()));
-            }
+            handle_error(tcp_stream.write_all(message.as_bytes()));
         });
 
         thread::spawn(move || {